--- Controls-demo/toggle/Tumbler/MainCounterStyle/MainCounterStyle.wml
@@ -1,17 +1,0 @@
-<div class="controlsDemo__wrapper {{_options.theme.indexOf('default') < 0 ? 'controlsDemo_fixedWidth500' : 'controlsDemo_fixedWidth300'}}">
-    <div class="controlsDemo__cell">
-        <div class="controls-text-label controlsDemo_widthFit controls-text-label">MainCounterStyle default</div>
-        <Controls.toggle:Tumbler items="{{_items}}" bind:selectedKey="_selectedKey">
-            <ws:itemTemplate>
-                <ws:partial template="Controls/toggle:tumblerItemCounterTemplate" scope="{{itemTemplate}}"/>
-            </ws:itemTemplate>
-        </Controls.toggle:Tumbler>
-
-        <div class="controls-text-label controlsDemo_widthFit controls-text-label">MainCounterStyle secondary</div>
-        <Controls.toggle:Tumbler items="{{_items}}" bind:selectedKey="_selectedKey">
-            <ws:itemTemplate>
-                <ws:partial template="Controls/toggle:tumblerItemCounterTemplate" scope="{{itemTemplate}}"/>
-            </ws:itemTemplate>
-        </Controls.toggle:Tumbler>
-    </div>
-</div>
